rename RemoveUnusedMemoryStage to ExploitDataLocalityStage
