+++ docs/source/stages.rst
@@ -119,8 +119,8 @@ Cost model stages
 
 Hardware modification stages
 -----------------
-* `SearchUnusedMemoryStage <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/SearchUnusedMemoryStage.py#L74>`_: Class that iterates through the memory instances and return the lowest allowed memory level for each operand for the usage of the next layer. The class must be placed before the WorkloadStage. The parameter `workload_data_always_from_top_mem <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/SearchUnusedMemoryStage.py#L164>`_ is False by default, which means the initial input and final output of the entire workload can be from a memory level lower than the highest memory level. You can set it to True if the initial input data and final output of the entire workload must travel from/to the highest memory level.
-* `RemoveUnusedMemoryStage <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/RemoveUnusedMemoryStage.py#L34>`_: Class that remove the unused memory instances according to the result of SearchUnusedMemoryStage. Each memory instance with a level higher than the level returned from SearchUnusedMemoryStage will be considered as an unused memory and will be removed. This stage must be placed after the WorkloadStage.
+* `SearchInterLayerDataLocalityStage <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/SearchInterLayerDataLocalityStage.py#L74>`_: Class that iterates through the memory instances and return the lowest allowed memory level for each operand for the usage of the next layer. The class must be placed before the WorkloadStage. The parameter `workload_data_always_from_top_mem <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/SearchInterLayerDataLocalityStage.py#L164>`_ is False by default, which means the initial input and final output of the entire workload can be from a memory level lower than the highest memory level. You can set it to True if the initial input data and final output of the entire workload must travel from/to the highest memory level.
+* `ExploitInterLayerDataLocalityStage <https://github.com/KULeuven-MICAS/zigzag/blob/master/zigzag/classes/stages/ExploitInterLayerDataLocalityStage.py#L34>`_: Class that remove the unused memory instances according to the result of SearchInterLayerDataLocalityStage. Each memory instance with a level higher than the level returned from SearchInterLayerDataLocalityStage will be considered as an unused memory and will be removed. This stage must be placed after the WorkloadStage.
 
 Creating your custom stage
 ==========================
