+++ tests/main/test_with_exploit_data_locality/test_edge_tpu_like.py
@@ -2,7 +2,7 @@ import pytest
 import sys
 
 sys.path.append("../zigzag")
-from zigzag.api import get_hardware_performance_zigzag_without_unused_memory
+from zigzag.api import get_hardware_performance_zigzag_with_exploit_data_locality
 
 workloads = (
     "zigzag/inputs/workload/resnet18.onnx",
@@ -11,26 +11,26 @@ workloads = (
 
 # Expected energy and latency for each workload defined above
 ens_lats = {
-    "zigzag/inputs/workload/alexnet.onnx": (6039993369, 8373281),
-    "zigzag/inputs/workload/mobilenetv2.onnx": (930699349, 1955657),
-    "zigzag/inputs/workload/resnet18.onnx": (1724869587, 3247768),
-    "zigzag/inputs/workload/resnet18.yaml": (2220200680, 3924484),
+    "zigzag/inputs/workload/alexnet.onnx": (5568602396.684999, 8134431),
+    "zigzag/inputs/workload/mobilenetv2.onnx": (735250234.1699998, 2417321.0),
+    "zigzag/inputs/workload/resnet18.onnx": (1783299827.71, 3156625.0),
+    "zigzag/inputs/workload/resnet18.yaml": (2115121959.8699996, 3855157.0),
 }
 
 
 @pytest.fixture
 def mapping():
-    return "zigzag/inputs/mapping/tesla_npu_like.yaml"
+    return "zigzag/inputs/mapping/edge_tpu_like.yaml"
 
 
 @pytest.fixture
 def accelerator():
-    return "zigzag/inputs/hardware/tesla_npu_like.yaml"
+    return "zigzag/inputs/hardware/edge_tpu_like.yaml"
 
 
 @pytest.mark.parametrize("workload", workloads)
 def test_api(workload: str, accelerator: str, mapping: str):  # pylint: disable=W0621
-    (energy, latency, _) = get_hardware_performance_zigzag_without_unused_memory(workload, accelerator, mapping)
+    (energy, latency, _) = get_hardware_performance_zigzag_with_exploit_data_locality(workload, accelerator, mapping)
     (expected_energy, expected_latency) = ens_lats[workload]
     print(f"'{workload}': ({energy}, {latency}),")
     assert energy == pytest.approx(expected_energy)  # type: ignore
