+++ tests/main/test_with_exploit_data_locality/test_meta_prototype_like.py
@@ -2,7 +2,7 @@ import pytest
 import sys
 
 sys.path.append("../zigzag")
-from zigzag.api import get_hardware_performance_zigzag_without_unused_memory
+from zigzag.api import get_hardware_performance_zigzag_with_exploit_data_locality
 
 workloads = (
     "zigzag/inputs/workload/resnet18.onnx",
@@ -30,7 +30,7 @@ def accelerator():
 
 @pytest.mark.parametrize("workload", workloads)
 def test_api(workload: str, accelerator: str, mapping: str):  # pylint: disable=W0621
-    (energy, latency, _) = get_hardware_performance_zigzag_without_unused_memory(workload, accelerator, mapping)
+    (energy, latency, _) = get_hardware_performance_zigzag_with_exploit_data_locality(workload, accelerator, mapping)
     (expected_energy, expected_latency) = ens_lats[workload]
     print(f"'{workload}': ({energy}, {latency}),")
     assert energy == pytest.approx(expected_energy)  # type: ignore
